If user creates a poll right in that time, instance will be automatically generated
